Remove networking module
Turns out we don't need this stuff.  Lambdas only need to live in a
VPC if they need to access resources that live in a VPC.  But the only
thing they need to access is DynamoDB, which doesn't live in a VPC.

We can add it in again if we do need it again in future.

When we merge this PR, we should manually `terraform destroy` the
project to clean it up.

See also alphagov/di-ipv-config#23