PYI-356: Set up dynamodb table for user credentials
Defines a table in DynamoDB for the lambdas to store and retrieve user
info received from credential issuers.

It defines it with a composite primary key based on the users session ID
and the credential issuer. This will have to change in future when we
use longer term IDs for users.

The TF also defines the IAM policy to allow the lambdas to access the
DB, as well as providing the table name to the lambda as an env var.