PYI-364: Update CredentialIssuerException
Adds a status code and error response field to the
CredentialIssuerException.

This allows the exception to be reused and have specific context when
catching errors in separate parts of the lambda. These are then returned
by the handler which uses them in it's response.

Is this a good idea? Should the service be telling the handler anything
about http status codes. We should talk about it...

Co-authored-by: Phil Miller <phillip.miller@digital.cabinet-office.gov.uk>
Co-authored-by: Sunil Dussoye <sunil.dussoye@digital.cabinet-office.gov.uk>