PYI-403: Use transformer with SSMProvider
The SSM provider from powertools has the ability to do the deserializing
for you. You just need to write a custom transformer.

This converts our current logic to a transformer for base64 encoded
yaml. This should mean we can use it in future if we need to deserialize
other yaml.