BAU: Test local SSM Client
This adds a test to ensure that the SSM client is sending requests to
localhost when the "IS_LOCAL" env variable is set to true.

This was mostly done to implement the mocking of environment variables.
This is achieved with the systems stubs library.

We depend on env variables more and more and they make testing
difficult. Having an easy way to mock them should be valuable.