BAU: Increase test coverage
This adds a few tests to cover the lines that are missing according to
sonarcloud.

A lot of these are just testing the no arg constructors for the handlers
and services used by the lambdas. Whilst there is not much merit in
testing them this way, it does boost our coverage report.

Our quality gate is set to 80% coverage. Without these changes we're
quite close to this boundary. Having code in these constructors drag us
down isn't helpful.