PYI-554: Add lambda to return all credentials
This is very similar to the useridentity lambda except without the token
check. This is because the request is coming from the same (our) system.

For now it returns some stub data, but can easily be updated to actually
chat to the database.

I've tried to import as few dependencies as possible, hence the use of
java.util.logging. Very much up for debate.

This doesn't deploy the thing.