PYI-526 Add error handling to CRI config parsing
If the credential issuer config from Parameter Store cannot be parsed into
CredentialIssuerConfig class because the id or attribute name cannot be split
from the parameter store path then throw and exception and return a 500 to the
client. This was thought preferable then logging and carrying on with the other
credential issuer config because other errors may occur within IPV core if the
config is incorrect.

with @danomanic and @sunilDGDS