PYI-612: Add shared attributes parameters
This renames the "credential-issuers" module to "parameters" to reflect
that it now contains the terraform for the credential issuers parameters
as well as other stuff.

It also adds some new parameters for the shared attributes signing key
and cert and adds a policy to the SAM template for the shared attributes
function.

It might the case that management of all ssm parameters will be moving
to SAM. In the interest of keeping things in the same place for now I've
defined these params in the terraform, rather than in SAM. This might
make the transition easier if it does happen. If it doesn't, then at
least we're being consistent.