PYI-640: Use new parameter store config structure
The layout of the config in the parameter store has changed. This
updates the templates to reflect it.

It also updates the configuration service to format the location of each
individual credential issuers config with the new format.

Some lambdas also had permissions on params they didn't need, which have
been removed.