PYI-680: Handle null values for redirect url's
It's valid in the spec for the redirect value to be omitted by the
client in the auth request. And if it is, it should also be omitted in
the token request.

At the moment this won't happen as we're validating the presence of the
redirect url in the auth endpoint, but we might handle it differently in
future.