PYI-764 and PYI-732: Get verifiable credentials from CRIs
CRIs will be returning verifiable credentials in the form of signed
JWTs. This updates the credentialissuerreturn lambda to send a post
request for the VC, and receive a signed JWT in response. It then
validates the signature on the response is expected.

We send a POST request to be in line with the Verifiable Credentials API
spec: https://w3c-ccg.github.io/vc-api/#introduction

We send the subject in an unsigned JWT to keep in line with this ADR.
Effectively the plain JWT is a "template" that the CRI can fill in. This
template will need to contain more information for back end only CRIs,
but that will get covered off in a future story. https://github.com/alphagov/digital-identity-architecture/blob/main/rfc/0008-identity-ipv-sequence-diagram.md#issue