PYIC-875: Return state with client response
The Oauth spec states that if a value for state is sent in the
authentication request, we should return in in the response. See here:
https://datatracker.ietf.org/doc/html/rfc6749#section-4.1.2

We currently validate the presence of the state value when we receive an
auth request from a client so we should always have one to return.
However, it is optional according to the spec, so this change will allow
for a missing value.

The auth teams client validates that the state received in the response
is the same as that sent in the initial auth request. We need to return
it to allow this to happen.