BAU: Convert der encoded signatures to concat
The nimbus lib we use to verify signatures on JWTs expects the
signatures to be in the "concat" format. Unfortunately the signatures
produced by AWS KMS are in the der format.

The nimbus lib provides a helper functions to convert between the two.
The concat version of an ES256 signature will always be 64 bytes long,
so we can leverage that to determine when to transcode signatures and
when to not.