PYIC-531: Run integration tests in Github actions
We want to run the integration tests as
part of our Github workflows. To do this Github actions will need to
talk to DynamoDB in the build account.

We already have a Github OIDC provider deployed in AWS which we use to
put build artifacts into S3. We can use the same patten to access
DynamoDB. A separate PR has been talk to AWS dynamodb, and so needs
extra permissions. raised on the config dir to give the AWS role perms
to DynamoDB.