PYIC-917: Give CRI return lambda SQS write permissions
For slice 2 we need to send an audit event from core-back when it
receives a VC and verifies its signature. It will send the audit event
to an SQS queue that has been created. This adds the permissions to
write to that queue.

We will need to also configure the lambda with the details of the SQS
queue - it's URL probably but maybe it's ARN. We will wait until we know
what the library that does the writing need before adding it.