PYIC-1091: Send audit event on vc recieve
Add an audit event `IPV_VC_RECEIVED` when a VC has been retrieved in the CredentialIssuerReturn lambda. This is in addition to the existing event `IPV_CREDENTIAL_RECEIVED_AND_SIGNATURE_CHECKED`. This is likely to be removed int he future, but we will send both for now.

At this stage we are only recording the event itself, with not payload. We will do that later when we have more details.