PYIC-992: Include enc key and audience in CRI config object
Add a credential issuers public encryption key and audience to the
CredentialIssuerConfig object.

We had these configuration details of CRIs defined separately. This is
most likely as they had been newly added. This refactors the object so
they are deserialized and available on the object with just one call the
parameter store.