PYIC-922: Fetch userId from session to validate VC JWT subject
We are storing the userID received from orch in the JAR they send us in
the session table. This is the value that's used as the subject when
making requests to CRIs. This is then passed back to us as the subject
of the VC JWT.

This pulls that value from the session and validates subject in the JWT
using it.

This also DRYs up the CRI return lambda tests a bunch.