PYIC-1251 Add a Private API Gateway
This adds a private API Gateway which can only be invoked via a VPC
Endpoint in the environment's VPC. This will replace the existing
"internal" core back API gateway which is public. In the interim this
new private API Gateway invokes the same lambdas on the same paths as
the existing public API Gateway.