PYIC-1251 Remove Open API Spec for Private API Gateway
A limition in SAMs API auth support means that `sam validate` will not
work when the API Gateway body definition is via a separate Open API
Spec referenced via an `AWS::Includes` transform. For now this removes
the `DefinitionBody` parameter so that CFN generates its own based on
the template contents. If we believe this needs enriching with our own
API Spec we can address that at a later date.