PYIC-1295: Validate KBV check
This adds logic to the validatecricheck lambda to allow us to determine
if a KBV check has reached a high enough confidence level to proceed
with an M1A journey.

I've gone for a different approach for how to organise the validators.
I've attempted to use composition as a technique to allow encapsulation
of the policies for different credential issuers, whilst allowing common
logic to be shared. Does this make things unnecessarily complicated or
easy to understand? You tell me...