PYIC-1296: Add safety checks in Evidence validator
The VCs we receive should always contain an evidence property that is an
array. But it's better to be safe than sorry and do a couple of checks
before trying to process it.

This catches a class cast exception in case the array is not an array,
and adds a null check in case the array does not exist.