PYIC-1296: Validate fraud check
This adds a new evidence validator to encapsulate the rules for
validating a fraud check and hooks it up to the CriCheck validator.

The ticket for this work specifically calls out that the identity fraud
score should be 1 or 2. Looking at GPG45, it's possible for this score
to also be 3: https://www.gov.uk/government/publications/identity-proofing-and-verification-of-an-individual/how-to-prove-and-verify-someones-identity#identity-fraud

I've set the rule to just ensure that the value is 1 or greater. I can't
see that we'd fail someone who received a score of three, even though
we'll never actually receive them.