PYIC-1377: Don't explicitly set expiry time on session
Rather than setting an expiry time on the session when it's created, it
will be better to check the creation time + timeout period at the point
we want to see if the session has expired. That way, any changes to the
timeout period will apply to all sessions, including those that already
exist.