PYIC-858: Return address attribute to orch
This extracts the address from the VC returned by the address CRI, and
adds it as a standalone claim in the JSON returned to the orchestrator
from the core. The spec for this can be found in RFC14: https://github.com/alphagov/digital-identity-architecture/blob/main/rfc/0014-identity-ipv-core-oauth-profile.md

It was decided that we would only return this data if we're at P2, and
that scoring would be added further down the line.

I've deliberately avoided deserializing the JSON contained in the
address VC into an Address domain object. Partly because it doesn't feel
like the core's responsibility to validate the structure at all at this
point in the journey, and partly because it means that things won't
break if the structure is ever updated in the address CRI. We just pass
along what we've been given. If people feel we'd be better using the
Address domain object we already have (updated in previous commit) I'm
happy to update it.