BAU: Parse token response as OAuth response
We were using an OIDC token response parser, rather than an OAuth token
response parser. While this still works, it also attempts to parse an
id_token as well as access_token and refresh_token. Using an OIDC thing
in a non OIDC flow was confusing, so we should switch to the correct
parser.