PYIC-1485: Log in JSON format
This leverages the AWS docs found here:
https://docs.aws.amazon.com/lambda/latest/dg/java-logging.html#java-logging-log4j2
as well as RFC0030:
https://github.com/alphagov/digital-identity-architecture/blob/main/rfc/0030-cross-programme-logging-format.md
and the auth teams previous good work:
https://github.com/alphagov/di-authentication-api/blob/main/ipv-api/src/main/resources/log4j2.xml

This uses the `aws-lambda-java-log4j2` library to allow is to include
the AWS request ID in each log line.

It defines a couple of new gradle configurations in the root project
which are then pulled in by the sub projects to reduce repetition.

It adds a YAML configuration file to each sub project to configure the
logging. Currently we're only adding a couple of custom fields,
session-id and aws-request-id. We can add more as we identify what they
are. The log4j2 docs are good for understanding what's happening:
https://logging.apache.org/log4j/2.x/manual/configuration.html#YAML
https://logging.apache.org/log4j/2.x/manual/json-template-layout.html

As we're already using slf4j to create our loggers we don't need to
update our code o get hold of a logger which is nice.