PYIC-1487: Add IDs to JSON logs
This updates each lambda to add the ipv session ID, the credential
issuer ID, the client ID and a component ID to the logs as soon as
they're available.

The component ID is a constant thing we add to help with filtering logs.
The same value will be added in the frontend logs. I've intentionally
left it environment agnostic (so not a URL) so it's easy to handle
across all deployments.

The logging makes use of the AWS Powertools logging lib. This has
simplified our dependencies, and adds some extra fields to our logs for
free. It also handles clean up of the MDC for us.

Instantiation of the loggers no longer uses slf4j. We're not going to
be changing the logging framework anytime soon and it makes things a
little simpler.

This change has also altered the responses of some of the lambdas in
some situations. Rather than returning a journey error response, I've
changed it to just returning a 400. This happens where we the
ipv-session-id has not been sent in the header. If this ever happens
something is seriously wrong and trying to continue the journey is
probably a bad idea and we should just throw. Opinions encouraged.