BAU: Parse auth grant instead of token request
When receiving an request to the access token endpoint we have been
manually recreating a TokenRequest object from an HTTPRequest and an
arbitrary URL. Rather than jumping through these hoops we can just parse
the auth grant and scope from the params directly.

This does raise the question of what are we doing with the requested
scope value. We have been playing back to the client whatever scope they
requested. We don't really use scopes in any meaningful way so this has
been fine, but we should probably be returning a default value, or no
value at all. This will get dealt with in another PR after talking to
the auth team about what they're sending us.