BAU: Fix structured logging of message
We want our logs to be structured to enable programatic analysis of them. Our previous attempt at this left the message field as a formatted string, rather than a correctly nested json object. We have extracted the Log4j2 `LambdaJsonLayout.json` from the Powertools logging library and updated it to allow MapMessages to be structured correctly. We then link to the modified version.

Co-authored-by: christopher.wynne@digital.cabinet-office.gov.uk