PYIC-1689: Determine if credentials meet M1A
This adds logic to the evaluate-gpg45-scores lambda to see if the
credentials gathered already are sufficient to meet the M1A profile. It
uses code already written, pulled in from here, for defining the
profiles and scores: https://github.com/alphagov/di-ipv-core-back/tree/wip/gpg45-lib/src/main/java/uk/gov/di/ipv/gpg45

When building the scores from the gathered credentials, for gpg45 scores
that only have one value (unlike evidence which can have multiple), this
will pick the score that has the highest value with the lowest number of
contra indicators. This will be useful in the future, but currently we
only ever expect to have score for each type of score so there is
actually nothing to sort.

If the credentials are determined to have met the M1A profile we end the
users session. If not, we continue them on the journey.