PYIC-1689: Fail journey in gpg45 evaluator if failing credential
We need to fail a journey if we receive a VC from a CRI that won't allow
us to meet M1A. Currently this is done by the validatecricheck lambda.

This steals some of the logic from there to do an upfront check on all
the credentials gathered so far in the evaluate-gpg45-scores lambda. It
will fail the journey if any VCs don't meet M1A.

This logic will be moved to the CRI selector lambda in future - this
lambda should just be responsible for deciding if we've met a profile or
if we need more evidence. The CRI selector should be responsible for
failing a journey if it decides there are no more paths for a user to
take.

Adding this in makes some of the logic added in the previous commit
redundant - specifically the handling of CIs. This logic will be needed
when the fail check is moved to the CRI selector lambda so it should
stay.