PYIC-1739: Delete existing VCs when starting session
As we now store VCs against the users ID, rather than there session, we
end up sending old names and addresses as part of the shared claims. For
example, if a user has tried a journey and made a mistake, then come
back later for another crack, we'll include the incorrect data from the
first attempt.

We do want to store VCs, but we need a proper think about how to handle
them. For now, we'll just delete any existing ones from the table when a
user starts a new session. This puts us back to a similar position to
how we were running before switching to user IDs as the primary key.