PYIC-1642: Update openAPI spec for private API
This updates the openAPI spec for the internal private API. This
includes endpoints that are shortly to be removed so will change very
soon. The important bit is the configuration for the
/journey/{journeyStep} endpoint. It sets up an integration with the new
step functions StartSyncExecution action.

It's necessary to do this as using SAM on its own to manage the
integration with the API gateway results in only the StartExecution
action being configured. With no way to change it. It also doesn't
appear to allow any form of mapping to be configured.

The mappings included here take an incoming HTTP request and parse out
the body, headers, path parameters and query parameters. It puts all
these values into a JSON object and sets it as the input for the state
machine. It also includes the ARN of the state machine we want to
trigger.

The response mapping just uses a default with a status code of 200. We
have to specify a status code here, but over ride it in the mapping
depending on what we get back from the state machine. The "status" being
checked here is the result of running the state machine. We can have a
"successful" execution that results in a 400, say, from the lambda. It's
only unsuccessful if there is an unexpected error. If the state machine
executed successfully, we pass back the output of the lambda and the
status code. Otherwise we return a 500.