PYIC-1642: Add StateMachine to the template
This adds the actual state machine resource to the deploy template. It
uses a very simple state machine definition that just passes all input
to the process-journey-step lambda and returns all of the output.

This also updates the internal API to use the updated openAPI spec.
There were some issues with `sam validate` not being able to handle the
use of the Auth block as well as the `DefinitionBody` block. There are
PRs for SAM to fix this, but in the mean time this includes a workaround
to allow the command to work.