PYIC-1642: Update process-journey-step lambda to use streams
Previously we were using a proxy integration with the lambdas and AWS
gateway. This meant that the incoming HTTP request was passed through to
the lambda. This could be modelled using the events library.

This no longer works as the input is very different. This updates the
lambda to use received its input from an input stream, and write it to
an output stream. There may well be better ways of doing this, but this
works for now.