PYIC-1864: Remove validate-cri-credential lambda
The validation of the gpg45 scores and CIs is now happening in the newly
wired in evaluate-gpg45-scores lambda, so we can remove the duplicated
logic.

We may bring this lambda back in future to do some other validation of
VC, such as signature checking. If we do it's probably easier to start
from fresh though.