BAU: Reinstate pre-kbv-transision page
In a recent refactor the pre-kbv transition page was skipped. This puts
it back in the flow.

It also updates the events for the build environment CRI_STATE state.
These should be stubs. The build env is currently broken.

I have no idea how any of this made it to prod...