PYIC-1834: Contra-indicator storage core stub
Adds a lambda that will be deployed to dev and build accounts that will
allow interaction with the contra-indicator storage system for dev and
testing purposes.

It provides a web page that can be used to submit CIs and a user ID to
the system, wrapped in a signed JWT. It will also retrieve the CIs for a
given user ID.

This makes an assumption of the interface for the getContraindicator
lambda, which has not been built yet, and so may need adjusting in
future.