PYIC-1834: Move CI storage core stub to own stack
This moves the CloudFormation to deploy the CI storage core stub to its
own stack. This means it can be deployed independently of the core,
which is a good thing. Less risk of it being deployed to an undesired
environment, and less complexity in the templates.

Two new secure pipelines have been created in the dev and build accounts
to do the actual deploy. This commit includes a new GitHub workflow to
build the stub and push the artifacts to those accounts for the
pipelines to pick up and deploy.