PYIC-1642: Embed process-journey-step lambda in Express Step Function
This re-applies the previously reverted work to embed the
process-journey-step lambda in a step function.

The original PR failed to deploy to staging for non obvious reasons and
was reverted to prevent blocking the pipeline.

This PR is to try again, and be able to investigate if it fails to
deploy again.

The original PR is here:
https://github.com/alphagov/di-ipv-core-back/pull/386