PYIX-1834: Deploy build stub to different bucket
The stub was failing to work in the build environment due to the
requirement of the secure-pipeline to include a permissions boundary for
the lambda. The permissions boundary does not allow cross account lambda
invoke.

We can get round this for now by using Concourse to deploy it to build.
This requires uploading the artifacts to a different bucket that
Concourse has permissions to read from. The bucket we're currently
deploying to is locked down to only be read by the secure pipeline.

This also changes the role we're assuming to do the deploying, as that
role has permissions on this bucket.

This is all temporary until the permissions boundary issue is resolved
(which it will need to be for core to use the new system), and then we
can go back to deploying how we were before.