PYIC-1864: Update debug journey for fail pages
Previously the validation of the gpg45 scores and contra indicators was
being done immediately after the user returned from a CRI. This meant
the debug journey flow worked with the "no-match" pages we show.

Now, we're evaluating the scores just before we send a user to a CRI.
This means that the debug flow misses the opportunity to show the
"no-match" pages before going back to the debug page.

This is breaking our smoke tests.

This updates the debug flow to emulate the acutal flow, by running the
gpg45 evaluator lambda before going returning to the debug page.

There's a good chance this will break other smoke tests, as it will
start showing the "no-match" pages after a visit to a stub CRI. This is
probably what we want though and the tests will need fixing up.