PYIC-1904: recordedAt -> issuedAt
The nbf claim of a JWT is confusingly mapped to the to the issuanceDate
property of the underlying VC.

Originally we were storing this date to assist with the two year ttl.
However, we’re using the ttl attribute in the DDB to manage that in
keeping with how we do this elsewhere.

We’ve decided to keep this timestamp for flexibility in the future, but
we should rename it to better reflect what it actually is.