BAU: Remove contra-indicator-storage-core-stub
The stub was created to allow us to exercise the CI storage system
independently. This was for automated testing purposes.

We've agreed for now to assume that core and the CI storage system are
connected in build. This means we can test the system with the existing
smoke tests.

If we decided we want this stub in future, we've got it in git history.