PYIC-1889: Process CIs from storage service
This adds logic to start using the CIs stored in the storage service to
generate a journey response. The result is currently just compared to
the existing logic and logged, rather than being acted upon. This can
stay this way until we're ready to start using the system in prod.

This approach uses the issuer of the last known CI to determine which
error page to show to a user. This should be consistent with current
behaviour, but may not be sustainable long term.

This new logic is also light on tests. When we actually come to start
using it, we can add more tests to give us better confidence in it.