PYIC-1865: Start storing CIs in prod
This removes the check on the environment and starts storing
contra-indicators in the production environment.

It also removes the error handling around the call to the new system.
This means that if there is an issue the users journey will be
interrupted.

Any VCs submitted to the CI storage system that result in an error will
get dropped, as per the RFC.