BAU: Handle VCs with multiple names
The way we were generating shared claims wasn't working when a VC
contained multiple names. The code was taking all the nameParts from all
the names and combining them into one name. This was then being
included as an extra name in the final shared claims.

This updates the logic to keep each name received in the VC as an
individual. All names from all VCs are then combined into a final set to
do the deduplication.