BAU: Remove Jackson dependency again
Jackson is already brought in by the AWS SDK so we don't need to define
it again.

Last time I tried this I broke all the application logs in production.
The reason for this was the log4j2 config was in yaml format, and
required the `jackson-dataformat-yaml` package to parse. Without it it
failed silently.

This changes the log4j2 config to JSON format. This only needs Jackson
core and databind, which are already brought in. This allows us to drop
all the Jackson dependencies.