PYIC-2054: Add expiration time to stored VCs
If we’re going to dump stored VCs that are expiring before the end of
the users current session, we’ll need to compare the exp claim in the VC
against the users session end time. We’ll need to do this for all their
VCs when they start a session.

Rather than having to pull all the VCs from Dynamo and then parse them,
we can just store the VCs expiration time as a DDB attribute when we
store them. We’ve already got it in it’s parsed form at that point so
this adds no extra computation, other than storing an extra field.

Once we have this, we can use Dynamo’s filtering capability to just
delete expiring VCs, without having to pull everything back, parse them,
compare the expiration time and then delete certain VCs.

This also changes `LocalDateTime` to Instant in the Bean class for
`UserIssuedCredentialsItem`. `LocalDateTime` doesn't represent a specific
point on the timeline - it has no timezone associated with it so
naturally has ambiguity. `LocalDateTime.now()` will give you a BST time.
We should be using UTC to prevent time weirdness. Instant is aligned
with UTC and so represents an actual moment in time.